+++ javascripts/discourse/connectors/topic-above-posts/ddi-document-timeline.js
@@ -1,0 +1,13 @@
+import { getOwner } from "@ember/owner";
+
+export default {
+  setupComponent(args, component) {
+    const metadata = getOwner(component)
+      .lookup("service:ddi-document-metadata")
+      .getMetadata(args.model);
+
+    component.setProperties({
+      timelineEvents: metadata.timeline,
+    });
+  },
+};